+++ hypervisor/arch/x86/guest/guest.c
@@ -34,7 +34,7 @@ is_vm0(struct vm *vm)
 	return (vm->attr.boot_idx & 0x7FU) == 0;
 }
 
-inline struct vcpu *vcpu_from_vid(struct vm *vm, int vcpu_id)
+inline struct vcpu *vcpu_from_vid(struct vm *vm, uint16_t vcpu_id)
 {
 	int i;
 	struct vcpu *vcpu;
